fix(hf): retry rate-limited tree listing before mutation
Retry only the idempotent Hugging Face list_repo_files preflight on HTTP 429 with bounded exponential backoff and Retry-After support. Preserve immediate failure for non-429 errors and keep uploads, commits, deletes, source binding, and restart outside the retry boundary.
