--- .github/workflows/zz-hf-tree-list-retry-429-once.yml
@@ -1,86 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-name: HF tree-list retry repair once
-
-on:
-  push:
-    branches:
-      - fix/hf-list-retry-429
-    paths:
-      - .github/scripts/repair_hf_tree_list_retry_429.py
-      - .github/workflows/zz-hf-tree-list-retry-429-once.yml
-
-permissions:
-  contents: write
-
-concurrency:
-  group: hf-tree-list-retry-429-once
-  cancel-in-progress: false
-
-jobs:
-  repair:
-    runs-on: ubuntu-latest
-    timeout-minutes: 15
-    steps:
-      - name: Harden runner
-        uses: step-security/harden-runner@05e31511f85b41b11d1cf0ef85d0992719546e2c # v2.21.0
-        with:
-          egress-policy: audit
-
-      - name: Checkout exact repair branch
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-        with:
-          ref: fix/hf-list-retry-429
-          fetch-depth: 0
-          persist-credentials: true
-
-      - name: Prove bounded execution lineage
-        shell: bash
-        env:
-          BASE_SHA: 4f916583a05c1208dcb34e0e4f2d78ccf3636ced
-        run: |
-          set -euo pipefail
-          git merge-base --is-ancestor "$BASE_SHA" HEAD
-          git diff --name-only "$BASE_SHA"...HEAD | sort > /tmp/actual-paths
-          printf '%s\n' \
-            '.github/scripts/repair_hf_tree_list_retry_429.py' \
-            '.github/workflows/zz-hf-tree-list-retry-429-once.yml' \
-            | sort > /tmp/expected-paths
-          diff -u /tmp/expected-paths /tmp/actual-paths
-
-      - name: Set up Python
-        uses: actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97 # v7.0.0
-        with:
-          python-version: '3.12'
-
-      - name: Apply deterministic repair and offline regressions
-        run: python3 .github/scripts/repair_hf_tree_list_retry_429.py
-
-      - name: Remove one-shot execution files and prove final scope
-        shell: bash
-        env:
-          BASE_SHA: 4f916583a05c1208dcb34e0e4f2d78ccf3636ced
-        run: |
-          set -euo pipefail
-          git rm \
-            .github/scripts/repair_hf_tree_list_retry_429.py \
-            .github/workflows/zz-hf-tree-list-retry-429-once.yml
-          git add -A
-          git diff --cached --check
-          git diff --name-only "$BASE_SHA" | sort > /tmp/final-paths
-          printf '%s\n' \
-            '.github/scripts/hf_deploy_from_dockerfile.py' \
-            '.github/scripts/test_hf_deploy_from_dockerfile.py' \
-            | sort > /tmp/expected-final-paths
-          diff -u /tmp/expected-final-paths /tmp/final-paths
-
-      - name: Commit and push exact repair
-        shell: bash
-        run: |
-          set -euo pipefail
-          git config user.name 'Lutar, Stephen P.'
-          git config user.email 'stephenlutar2@gmail.com'
-          git commit \
-            -m 'fix(hf): retry rate-limited tree listing before mutation' \
-            -m 'Retry only the idempotent Hugging Face list_repo_files preflight on HTTP 429 with bounded exponential backoff and Retry-After support. Preserve immediate failure for non-429 errors and keep uploads, commits, deletes, source binding, and restart outside the retry boundary.' \
-            -m 'Signed-off-by: Stephen Lutar <stephenlutar2@gmail.com>'
-          git push origin HEAD:fix/hf-list-retry-429
